+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- chore(dependencies): update @harborclient/sdk to version 1.0.72. (`d79aa21`)
 - refactor(terminal-panel): remove description from TerminalPanel component. (`00a0d99`)
 
 ## 2.2.0 - 2026-07-12